26/12/24 - Modularity & refined midpoint values.
Broke down sections of repeated code (into a Custom Functions section) into the following functions:

Removing records with duplicate DateTimes from a given dataset.

Calculate the midpoints of the statistical summary DataFrame with a given rolling window size. Midpoint is rounded up to the nearest whole number to match index of the midpoint to mimic the rolling window process including adjusting the midpoint calculation and index positions at the ends of the DataFrame to not go out of bounds.

Functions to plot out each of the statistical summary graphs given the statistical summary DataFrame and rolling window size in order to not rewrite the same processes repeatedly.

Corrected miscellaneous syntax errors through misspelt words and incorrect keyworks when cells were run.